requirements.txt
streamlit==1.41.1
pandas==2.2.3
requests==2.32.3
folium==0.17.0
streamlit-folium==0.23.2
openpyxl==3.1.5
reportlab==4.2.5
PyPDF2==3.0.1
PyMuPDF>=1.23.0
Pillow>=10.0.0
numpy>=1.24.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
opencv-python>=4.8.0
